+++ tests/app/opulence/console/responses/compilers/CompilerTest.php
@@ -6,10 +6,9 @@
  */
 namespace Opulence\Console\Responses\Compilers;
 
+use Opulence\Console\Responses\Compilers\Elements\Style;
 use Opulence\Console\Responses\Compilers\Lexers\Lexer;
 use Opulence\Console\Responses\Compilers\Parsers\Parser;
-use Opulence\Console\Responses\Formatters\Elements\Element;
-use Opulence\Console\Responses\Formatters\Elements\Style;
 use RuntimeException;
 
 class CompilerTest extends \PHPUnit_Framework_TestCase
@@ -30,8 +29,8 @@ class CompilerTest extends \PHPUnit_Framework_TestCase
      */
     public function testCompilingAdjacentElements()
     {
-        $this->compiler->getElements()->add(new Element("foo", new Style("green", "white")));
-        $this->compiler->getElements()->add(new Element("bar", new Style("cyan")));
+        $this->compiler->registerElement("foo", new Style("green", "white"));
+        $this->compiler->registerElement("bar", new Style("cyan"));
         $expectedOutput = "\033[32;47mbaz\033[39;49m\033[36mblah\033[39m";
         $this->assertEquals(
             $expectedOutput,
@@ -44,8 +43,8 @@ class CompilerTest extends \PHPUnit_Framework_TestCase
      */
     public function testCompilingElementWithNoChildren()
     {
-        $this->compiler->getElements()->add(new Element("foo", new Style("green", "white")));
-        $this->compiler->getElements()->add(new Element("bar", new Style("cyan")));
+        $this->compiler->registerElement("foo", new Style("green", "white"));
+        $this->compiler->registerElement("bar", new Style("cyan"));
         $expectedOutput = "";
         $this->assertEquals(
             $expectedOutput,
@@ -59,8 +58,8 @@ class CompilerTest extends \PHPUnit_Framework_TestCase
     public function testCompilingElementWithoutApplyingStyles()
     {
         $this->compiler->setStyled(false);
-        $this->compiler->getElements()->add(new Element("foo", new Style("green", "white")));
-        $this->compiler->getElements()->add(new Element("bar", new Style("cyan")));
+        $this->compiler->registerElement("foo", new Style("green", "white"));
+        $this->compiler->registerElement("bar", new Style("cyan"));
         $this->assertEquals("bazblah", $this->compiler->compile("<foo>baz</foo><bar>blah</bar>"));
     }
 
@@ -69,7 +68,7 @@ class CompilerTest extends \PHPUnit_Framework_TestCase
      */
     public function testCompilingEscapedTagAtBeginning()
     {
-        $this->compiler->getElements()->add(new Element("foo", new Style("green")));
+        $this->compiler->registerElement("foo", new Style("green"));
         $expectedOutput = "<bar>";
         $this->assertEquals($expectedOutput, $this->compiler->compile("\\<bar>"));
     }
@@ -79,7 +78,7 @@ class CompilerTest extends \PHPUnit_Framework_TestCase
      */
     public function testCompilingEscapedTagInBetweenTags()
     {
-        $this->compiler->getElements()->add(new Element("foo", new Style("green")));
+        $this->compiler->registerElement("foo", new Style("green"));
         $expectedOutput = "\033[32m<bar>\033[39m";
         $this->assertEquals($expectedOutput, $this->compiler->compile("<foo>\\<bar></foo>"));
     }
@@ -89,8 +88,8 @@ class CompilerTest extends \PHPUnit_Framework_TestCase
      */
     public function testCompilingNestedElements()
     {
-        $this->compiler->getElements()->add(new Element("foo", new Style("green", "white")));
-        $this->compiler->getElements()->add(new Element("bar", new Style("cyan")));
+        $this->compiler->registerElement("foo", new Style("green", "white"));
+        $this->compiler->registerElement("bar", new Style("cyan"));
         $expectedOutput = "\033[32;47m\033[36mbaz\033[39m\033[39;49m";
         $this->assertEquals(
             $expectedOutput,
@@ -103,8 +102,8 @@ class CompilerTest extends \PHPUnit_Framework_TestCase
      */
     public function testCompilingNestedElementsWithNoChildren()
     {
-        $this->compiler->getElements()->add(new Element("foo", new Style("green", "white")));
-        $this->compiler->getElements()->add(new Element("bar", new Style("cyan")));
+        $this->compiler->registerElement("foo", new Style("green", "white"));
+        $this->compiler->registerElement("bar", new Style("cyan"));
         $expectedOutput = "";
         $this->assertEquals(
             $expectedOutput,
@@ -117,8 +116,8 @@ class CompilerTest extends \PHPUnit_Framework_TestCase
      */
     public function testCompilingNestedElementsWithWordsInBetween()
     {
-        $this->compiler->getElements()->add(new Element("foo", new Style("green", "white")));
-        $this->compiler->getElements()->add(new Element("bar", new Style("cyan")));
+        $this->compiler->registerElement("foo", new Style("green", "white"));
+        $this->compiler->registerElement("bar", new Style("cyan"));
         $expectedOutput = "\033[32;47mbar\033[39;49m\033[32;47m\033[36mblah\033[39m\033[39;49m\033[32;47mbaz\033[39;49m";
         $this->assertEquals(
             $expectedOutput,
@@ -143,7 +142,7 @@ class CompilerTest extends \PHPUnit_Framework_TestCase
      */
     public function testCompilingSingleElement()
     {
-        $this->compiler->getElements()->add(new Element("foo", new Style("green")));
+        $this->compiler->registerElement("foo", new Style("green"));
         $expectedOutput = "\033[32mbar\033[39m";
         $this->assertEquals($expectedOutput, $this->compiler->compile("<foo>bar</foo>"));
     }
@@ -172,8 +171,8 @@ class CompilerTest extends \PHPUnit_Framework_TestCase
     public function testIncorrectlyNestedElements()
     {
         $this->setExpectedException(RuntimeException::class);
-        $this->compiler->getElements()->add(new Element("foo", new Style("green")));
-        $this->compiler->getElements()->add(new Element("bar", new Style("blue")));
+        $this->compiler->registerElement("foo", new Style("green"));
+        $this->compiler->registerElement("bar", new Style("blue"));
         $this->compiler->compile("<foo><bar>blah</foo></bar>");
     }
 }